Добавит проверку стоимости
